new file:   GUI_interface/GUI_interface.pde 	renamed:    vehicle-directory/3.txt -> GUI_interface/boot_screen.pde 	new file:   GUI_interface/data/gui/GUI_interface/saves/auto.json 	renamed:    vehicle-directory/809070.txt -> GUI_interface/diag_screen.pde 	new file:   GUI_interface/pdf_print.pde 	new file:   GUI_interface/preven_screen.pde 	modified:   diag.h 	modified:   diagnostic.cpp 	modified:   headers.h 	modified:   main.cpp 	modified:   preventative.cpp 	new file:   sketch.properties 	new file:   thing.pde 	new file:   univ-function.cpp 	deleted:    vehicle-directory/123 	modified:   vehicle-directory/123.txt 	deleted:    vehicle-directory/324 	deleted:    vehicle-directory/421 	deleted:    vehicle-directory/4231.txt 	deleted:    vehicle-directory/4231.txt4213.txt 	deleted:    vehicle-directory/hello
