+++ GUI_interface/GUI_interface.pde
@@ -1,0 +1,13 @@
+import com.krab.lazy.*;
+import processing.pdf.*;
+
+LazyGui gui;
+
+void setup(){
+    size(800,800,P2D);
+    gui = new LazyGui(this);
+}
+
+void draw(){
+    background(gui.colorPicker("background").hex);
+}
