+++ sketch.properties
@@ -1,0 +1,1 @@
+main=thing.pde
