--- vehicle-directory/123
@@ -1,13 +1,0 @@
-Be aware: the engine is flooded.
-Issue: faulty fuel pressure regulator.     Order: fuel pressure regulator.
-Diagnostic report for 123 vehicle:
-Issue reported: Engine
-Be aware: the engine is flooded.
-Issue: flooded engine.     Order: spark plugs, air filter.
-Diagnostic report for 123 vehicle:
-Issue reported: Engine
-Diagnostic report for 123 vehicle:
-Issue reported: Bodywork/Doors
-Issue: Rear driver's side door lock is non-functional. Order: send to third party to fix.
-Diagnostic report for 123 vehicle:
-Issue reported: Bodywork/Doors
