--- vehicle-directory/324
@@ -1,4 +1,0 @@
-Be aware: the engine is flooded.
-Issue unknown, third-party inspection required.
-Diagnostic report for 324 vehicle:
-Issue reported: Engine
