--- vehicle-directory/421
@@ -1,4 +1,0 @@
-Be aware: battery problems.
-Issue: weak or dead battery. Order: new battery.
-Diagnostic report for 421 vehicle:
-Issue reported: Engine Accessories
