--- vehicle-directory/4231.txt
@@ -1,14 +1,0 @@
-Preventative Maintenace Check for vehicle 4231:
-You need to replace the brake pads soon.
-Fix if you want.
-Battery may be low. Jump start the car.
-Your suspension is too old, replace suspension.
-Check and replace brake pads.
-Be aware: there are transmission fluid leaks.
-Issue: transmission cooler line leaking. Order: cooler lines.
-Diagnostic report for 4231 vehicle:
-Issue reported: Transmission
-Be aware: there are transmission fluid leaks.
-Issue unknown, third-party inspection required.
-Diagnostic report for 4231 vehicle:
-Issue reported: Transmission
